+++ Assets/Mirage/Components/HeadlessAutoStart.cs
@@ -24,7 +24,7 @@ namespace Mirage
             // some transports might not be ready until Start.
             if (Server && SystemInfo.graphicsDeviceType == GraphicsDeviceType.Null && startOnHeadless)
             {
-                Server.ListenAsync().Forget();
+                Server.StartAsync().Forget();
             }
         }
     }