+++ Assets/Tests/Runtime/Host/HostSetup.cs
@@ -23,6 +23,8 @@ namespace Mirage.Tests.Runtime.Host
         protected NetworkIdentity identity;
         protected T component;
 
+        protected virtual bool AutoStartServer => true;
+
         public virtual void ExtraSetup() { }
 
         [UnitySetUp]
@@ -53,16 +55,19 @@ namespace Mirage.Tests.Runtime.Host
             // wait for all Start() methods to get invoked
             await UniTask.DelayFrame(1);
 
-            await StartHost();
+            if (AutoStartServer)
+            {
+                await StartHost();
 
-            playerGO = new GameObject("playerGO", typeof(Rigidbody));
-            identity = playerGO.AddComponent<NetworkIdentity>();
-            component = playerGO.AddComponent<T>();
+                playerGO = new GameObject("playerGO", typeof(Rigidbody));
+                identity = playerGO.AddComponent<NetworkIdentity>();
+                component = playerGO.AddComponent<T>();
 
-            serverObjectManager.AddCharacter(server.LocalPlayer, playerGO);
+                serverObjectManager.AddCharacter(server.LocalPlayer, playerGO);
 
-            // wait for client to spawn it
-            await AsyncUtil.WaitUntilWithTimeout(() => client.Player.Identity != null);
+                // wait for client to spawn it
+                await AsyncUtil.WaitUntilWithTimeout(() => client.Player.Identity != null);
+            }
         });
 
         protected async UniTask StartHost()
@@ -76,7 +81,7 @@ namespace Mirage.Tests.Runtime.Host
 
             server.Started.AddListener(Started);
             // now start the host
-            manager.Server.StartHost(client).Forget();
+            manager.Server.StartAsync(client).Forget();
 
             await completionSource.Task;
         }
@@ -87,7 +92,7 @@ namespace Mirage.Tests.Runtime.Host
         public IEnumerator ShutdownHost() => UniTask.ToCoroutine(async () =>
         {
             Object.Destroy(playerGO);
-            manager.Server.StopHost();
+            manager.Server.Stop();
 
             await UniTask.Delay(1);
             Object.Destroy(networkManagerGo);